+++ app/services/issuer/acapy_issuer_v1.py
@@ -2,29 +2,27 @@ from typing import Dict, Optional
 
 from aries_cloudcontroller import (
     AcaPyClient,
-    V20CredAttrSpec,
-    V20CredExFree,
-    V20CredExRecord,
-    V20CredFilter,
-    V20CredFilterIndy,
-    V20CredOfferConnFreeRequest,
-    V20CredPreview,
-    V20CredRequestRequest,
+    CredAttrSpec,
+    CredentialPreview,
+    V10CredentialConnFreeOfferRequest,
+    V10CredentialExchange,
+    V10CredentialProposalRequestMand,
 )
-from aries_cloudcontroller.model.v20_cred_store_request import V20CredStoreRequest
+from aries_cloudcontroller.model.v10_credential_store_request import (
+    V10CredentialStoreRequest,
+)
 
-from app.exceptions.cloud_api_error import CloudApiException
-from app.facades.issuer.acapy_issuer import Issuer
+from app.services.issuer.acapy_issuer import Issuer
 from app.models.issuer import Credential, CredentialNoConnection
 from app.util.credentials import cred_id_no_version
 from shared.log_config import get_logger
-from shared.models.conversion import credential_record_to_model_v2
+from shared.models.conversion import credential_record_to_model_v1
 from shared.models.topics import CredentialExchange
 
 logger = get_logger(__name__)
 
 
-class IssuerV2(Issuer):
+class IssuerV1(Issuer):
     @classmethod
     async def send_credential(cls, controller: AcaPyClient, credential: Credential):
         bound_logger = logger.bind(body=credential)
@@ -33,20 +31,16 @@ class IssuerV2(Issuer):
             attributes=credential.attributes
         )
 
-        bound_logger.debug("Issue v2 credential (automated)")
-        record = await controller.issue_credential_v2_0.issue_credential_automated(
-            body=V20CredExFree(
+        bound_logger.debug("Issue v1 credential (automated)")
+        record = await controller.issue_credential_v1_0.issue_credential_automated(
+            body=V10CredentialProposalRequestMand(
                 connection_id=credential.connection_id,
-                credential_preview=credential_preview,
-                filter=V20CredFilter(
-                    indy=V20CredFilterIndy(
-                        cred_def_id=credential.cred_def_id,
-                    )
-                ),
+                credential_proposal=credential_preview,
+                cred_def_id=credential.cred_def_id,
             )
         )
 
-        bound_logger.debug("Returning v2 credential result as CredentialExchange.")
+        bound_logger.debug("Returning v1 credential result as CredentialExchange.")
         return cls.__record_to_model(record)
 
     @classmethod
@@ -59,21 +53,15 @@ class IssuerV2(Issuer):
             attributes=credential.attributes
         )
 
-        bound_logger.debug("Creating v2 credential offer")
-        record = (
-            await controller.issue_credential_v2_0.issue_credential20_create_offer_post(
-                body=V20CredOfferConnFreeRequest(
-                    credential_preview=credential_preview,
-                    filter=V20CredFilter(
-                        indy=V20CredFilterIndy(
-                            cred_def_id=credential.cred_def_id,
-                        )
-                    ),
-                )
+        bound_logger.debug("Creating v1 credential offer")
+        record = await controller.issue_credential_v1_0.create_offer(
+            body=V10CredentialConnFreeOfferRequest(
+                credential_preview=credential_preview,
+                cred_def_id=credential.cred_def_id,
             )
         )
 
-        bound_logger.debug("Returning v2 create offer result as CredentialExchange.")
+        bound_logger.debug("Returning v1 create offer result as CredentialExchange.")
         return cls.__record_to_model(record)
 
     @classmethod
@@ -86,12 +74,12 @@ class IssuerV2(Issuer):
         bound_logger.debug("Get credential id without version")
         credential_exchange_id = cred_id_no_version(credential_exchange_id)
 
-        bound_logger.debug("Sending v2 credential request")
-        record = await controller.issue_credential_v2_0.send_request(
-            cred_ex_id=credential_exchange_id, body=V20CredRequestRequest()
+        bound_logger.debug("Sending v1 credential request")
+        record = await controller.issue_credential_v1_0.send_request(
+            cred_ex_id=credential_exchange_id
         )
 
-        bound_logger.debug("Returning v2 send request result as CredentialExchange.")
+        bound_logger.debug("Returning v1 send request result as CredentialExchange.")
         return cls.__record_to_model(record)
 
     @classmethod
@@ -104,18 +92,15 @@ class IssuerV2(Issuer):
         bound_logger.debug("Get credential id without version")
         credential_exchange_id = cred_id_no_version(credential_exchange_id)
 
-        bound_logger.debug("Storing v2 credential record")
-        record = await controller.issue_credential_v2_0.store_credential(
-            cred_ex_id=credential_exchange_id, body=V20CredStoreRequest()
+        bound_logger.debug("Storing v1 credential record")
+        record = await controller.issue_credential_v1_0.store_credential(
+            cred_ex_id=credential_exchange_id, body=V10CredentialStoreRequest()
         )
-
-        if not record.cred_ex_record:
-            raise CloudApiException("Stored record has no credential exchange record.")
 
         bound_logger.debug(
-            "Returning v2 store credential result as CredentialExchange."
+            "Returning v1 store credential result as CredentialExchange."
         )
-        return cls.__record_to_model(record.cred_ex_record)
+        return cls.__record_to_model(record)
 
     @classmethod
     async def delete_credential(
@@ -127,21 +112,21 @@ class IssuerV2(Issuer):
         bound_logger.debug("Get credential id without version")
         credential_exchange_id = cred_id_no_version(credential_exchange_id)
 
-        bound_logger.debug("Getting v2 credential record")
-        record = await controller.issue_credential_v2_0.get_record(
+        bound_logger.debug("Getting v1 credential record")
+        record = await controller.issue_credential_v1_0.get_record(
             cred_ex_id=credential_exchange_id
         )
 
-        bound_logger.debug("Deleting v2 credential record")
-        await controller.issue_credential_v2_0.delete_record(
+        bound_logger.debug("Deleting v1 credential record")
+        await controller.issue_credential_v1_0.delete_record(
             cred_ex_id=credential_exchange_id
         )
 
         # also delete indy credential
-        if record.indy and record.indy.cred_id_stored:
+        if record.credential_id:
             bound_logger.debug("Deleting indy credential")
             await controller.credentials.delete_record(
-                credential_id=record.indy.cred_id_stored
+                credential_id=record.credential_id
             )
         bound_logger.debug("Successfully deleted credential.")
 
@@ -150,21 +135,17 @@ class IssuerV2(Issuer):
         cls, controller: AcaPyClient, connection_id: Optional[str] = None
     ):
         bound_logger = logger.bind(body={"connection_id": connection_id})
-        bound_logger.debug("Getting v2 credential records by connection id")
-        result = await controller.issue_credential_v2_0.get_records(
+        bound_logger.debug("Getting v1 credential records by connection id")
+        result = await controller.issue_credential_v1_0.get_records(
             connection_id=connection_id,
         )
 
         if not result.results:
-            bound_logger.debug("No v2 record results.")
+            bound_logger.debug("No v1 record results.")
             return []
 
-        bound_logger.debug("Returning v2 record results as CredentialExchange.")
-        return [
-            cls.__record_to_model(record.cred_ex_record)
-            for record in result.results
-            if record.cred_ex_record
-        ]
+        bound_logger.debug("Returning v1 record results as CredentialExchange.")
+        return [cls.__record_to_model(record) for record in result.results]
 
     @classmethod
     async def get_record(cls, controller: AcaPyClient, credential_exchange_id: str):
@@ -174,29 +155,26 @@ class IssuerV2(Issuer):
         bound_logger.debug("Get credential id without version")
         credential_exchange_id = cred_id_no_version(credential_exchange_id)
 
-        bound_logger.debug("Getting v2 credential record")
-        record = await controller.issue_credential_v2_0.get_record(
+        bound_logger.debug("Getting v1 credential record")
+        record = await controller.issue_credential_v1_0.get_record(
             cred_ex_id=credential_exchange_id,
         )
 
-        if not record.cred_ex_record:
-            raise CloudApiException("Record has no credential exchange record.")
-
-        bound_logger.debug("Returning v2 credential record as CredentialExchange.")
-        return cls.__record_to_model(record.cred_ex_record)
+        bound_logger.debug("Returning v1 credential record as CredentialExchange.")
+        return cls.__record_to_model(record)
 
     @classmethod
-    def __record_to_model(cls, record: V20CredExRecord) -> CredentialExchange:
-        return credential_record_to_model_v2(record=record)
+    def __record_to_model(cls, record: V10CredentialExchange) -> CredentialExchange:
+        return credential_record_to_model_v1(record)
 
     @classmethod
     def __preview_from_attributes(
         cls,
         attributes: Dict[str, str],
-    ) -> V20CredPreview:
-        return V20CredPreview(
+    ) -> CredentialPreview:
+        return CredentialPreview(
             attributes=[
-                V20CredAttrSpec(name=name, value=value)
+                CredAttrSpec(name=name, value=value)
                 for name, value in attributes.items()
             ]
         )
