+++ app/services/issuer/acapy_issuer_v2.py
@@ -2,27 +2,29 @@ from typing import Dict, Optional
 
 from aries_cloudcontroller import (
     AcaPyClient,
-    CredAttrSpec,
-    CredentialPreview,
-    V10CredentialConnFreeOfferRequest,
-    V10CredentialExchange,
-    V10CredentialProposalRequestMand,
+    V20CredAttrSpec,
+    V20CredExFree,
+    V20CredExRecord,
+    V20CredFilter,
+    V20CredFilterIndy,
+    V20CredOfferConnFreeRequest,
+    V20CredPreview,
+    V20CredRequestRequest,
 )
-from aries_cloudcontroller.model.v10_credential_store_request import (
-    V10CredentialStoreRequest,
-)
+from aries_cloudcontroller.model.v20_cred_store_request import V20CredStoreRequest
 
-from app.facades.issuer.acapy_issuer import Issuer
+from app.exceptions.cloud_api_error import CloudApiException
+from app.services.issuer.acapy_issuer import Issuer
 from app.models.issuer import Credential, CredentialNoConnection
 from app.util.credentials import cred_id_no_version
 from shared.log_config import get_logger
-from shared.models.conversion import credential_record_to_model_v1
+from shared.models.conversion import credential_record_to_model_v2
 from shared.models.topics import CredentialExchange
 
 logger = get_logger(__name__)
 
 
-class IssuerV1(Issuer):
+class IssuerV2(Issuer):
     @classmethod
     async def send_credential(cls, controller: AcaPyClient, credential: Credential):
         bound_logger = logger.bind(body=credential)
@@ -31,16 +33,20 @@ class IssuerV1(Issuer):
             attributes=credential.attributes
         )
 
-        bound_logger.debug("Issue v1 credential (automated)")
-        record = await controller.issue_credential_v1_0.issue_credential_automated(
-            body=V10CredentialProposalRequestMand(
+        bound_logger.debug("Issue v2 credential (automated)")
+        record = await controller.issue_credential_v2_0.issue_credential_automated(
+            body=V20CredExFree(
                 connection_id=credential.connection_id,
-                credential_proposal=credential_preview,
-                cred_def_id=credential.cred_def_id,
+                credential_preview=credential_preview,
+                filter=V20CredFilter(
+                    indy=V20CredFilterIndy(
+                        cred_def_id=credential.cred_def_id,
+                    )
+                ),
             )
         )
 
-        bound_logger.debug("Returning v1 credential result as CredentialExchange.")
+        bound_logger.debug("Returning v2 credential result as CredentialExchange.")
         return cls.__record_to_model(record)
 
     @classmethod
@@ -53,15 +59,21 @@ class IssuerV1(Issuer):
             attributes=credential.attributes
         )
 
-        bound_logger.debug("Creating v1 credential offer")
-        record = await controller.issue_credential_v1_0.create_offer(
-            body=V10CredentialConnFreeOfferRequest(
-                credential_preview=credential_preview,
-                cred_def_id=credential.cred_def_id,
+        bound_logger.debug("Creating v2 credential offer")
+        record = (
+            await controller.issue_credential_v2_0.issue_credential20_create_offer_post(
+                body=V20CredOfferConnFreeRequest(
+                    credential_preview=credential_preview,
+                    filter=V20CredFilter(
+                        indy=V20CredFilterIndy(
+                            cred_def_id=credential.cred_def_id,
+                        )
+                    ),
+                )
             )
         )
 
-        bound_logger.debug("Returning v1 create offer result as CredentialExchange.")
+        bound_logger.debug("Returning v2 create offer result as CredentialExchange.")
         return cls.__record_to_model(record)
 
     @classmethod
@@ -74,12 +86,12 @@ class IssuerV1(Issuer):
         bound_logger.debug("Get credential id without version")
         credential_exchange_id = cred_id_no_version(credential_exchange_id)
 
-        bound_logger.debug("Sending v1 credential request")
-        record = await controller.issue_credential_v1_0.send_request(
-            cred_ex_id=credential_exchange_id
+        bound_logger.debug("Sending v2 credential request")
+        record = await controller.issue_credential_v2_0.send_request(
+            cred_ex_id=credential_exchange_id, body=V20CredRequestRequest()
         )
 
-        bound_logger.debug("Returning v1 send request result as CredentialExchange.")
+        bound_logger.debug("Returning v2 send request result as CredentialExchange.")
         return cls.__record_to_model(record)
 
     @classmethod
@@ -92,15 +104,18 @@ class IssuerV1(Issuer):
         bound_logger.debug("Get credential id without version")
         credential_exchange_id = cred_id_no_version(credential_exchange_id)
 
-        bound_logger.debug("Storing v1 credential record")
-        record = await controller.issue_credential_v1_0.store_credential(
-            cred_ex_id=credential_exchange_id, body=V10CredentialStoreRequest()
+        bound_logger.debug("Storing v2 credential record")
+        record = await controller.issue_credential_v2_0.store_credential(
+            cred_ex_id=credential_exchange_id, body=V20CredStoreRequest()
         )
 
+        if not record.cred_ex_record:
+            raise CloudApiException("Stored record has no credential exchange record.")
+
         bound_logger.debug(
-            "Returning v1 store credential result as CredentialExchange."
+            "Returning v2 store credential result as CredentialExchange."
         )
-        return cls.__record_to_model(record)
+        return cls.__record_to_model(record.cred_ex_record)
 
     @classmethod
     async def delete_credential(
@@ -112,21 +127,21 @@ class IssuerV1(Issuer):
         bound_logger.debug("Get credential id without version")
         credential_exchange_id = cred_id_no_version(credential_exchange_id)
 
-        bound_logger.debug("Getting v1 credential record")
-        record = await controller.issue_credential_v1_0.get_record(
+        bound_logger.debug("Getting v2 credential record")
+        record = await controller.issue_credential_v2_0.get_record(
             cred_ex_id=credential_exchange_id
         )
 
-        bound_logger.debug("Deleting v1 credential record")
-        await controller.issue_credential_v1_0.delete_record(
+        bound_logger.debug("Deleting v2 credential record")
+        await controller.issue_credential_v2_0.delete_record(
             cred_ex_id=credential_exchange_id
         )
 
         # also delete indy credential
-        if record.credential_id:
+        if record.indy and record.indy.cred_id_stored:
             bound_logger.debug("Deleting indy credential")
             await controller.credentials.delete_record(
-                credential_id=record.credential_id
+                credential_id=record.indy.cred_id_stored
             )
         bound_logger.debug("Successfully deleted credential.")
 
@@ -135,17 +150,21 @@ class IssuerV1(Issuer):
         cls, controller: AcaPyClient, connection_id: Optional[str] = None
     ):
         bound_logger = logger.bind(body={"connection_id": connection_id})
-        bound_logger.debug("Getting v1 credential records by connection id")
-        result = await controller.issue_credential_v1_0.get_records(
+        bound_logger.debug("Getting v2 credential records by connection id")
+        result = await controller.issue_credential_v2_0.get_records(
             connection_id=connection_id,
         )
 
         if not result.results:
-            bound_logger.debug("No v1 record results.")
+            bound_logger.debug("No v2 record results.")
             return []
 
-        bound_logger.debug("Returning v1 record results as CredentialExchange.")
-        return [cls.__record_to_model(record) for record in result.results]
+        bound_logger.debug("Returning v2 record results as CredentialExchange.")
+        return [
+            cls.__record_to_model(record.cred_ex_record)
+            for record in result.results
+            if record.cred_ex_record
+        ]
 
     @classmethod
     async def get_record(cls, controller: AcaPyClient, credential_exchange_id: str):
@@ -155,26 +174,29 @@ class IssuerV1(Issuer):
         bound_logger.debug("Get credential id without version")
         credential_exchange_id = cred_id_no_version(credential_exchange_id)
 
-        bound_logger.debug("Getting v1 credential record")
-        record = await controller.issue_credential_v1_0.get_record(
+        bound_logger.debug("Getting v2 credential record")
+        record = await controller.issue_credential_v2_0.get_record(
             cred_ex_id=credential_exchange_id,
         )
 
-        bound_logger.debug("Returning v1 credential record as CredentialExchange.")
-        return cls.__record_to_model(record)
+        if not record.cred_ex_record:
+            raise CloudApiException("Record has no credential exchange record.")
+
+        bound_logger.debug("Returning v2 credential record as CredentialExchange.")
+        return cls.__record_to_model(record.cred_ex_record)
 
     @classmethod
-    def __record_to_model(cls, record: V10CredentialExchange) -> CredentialExchange:
-        return credential_record_to_model_v1(record)
+    def __record_to_model(cls, record: V20CredExRecord) -> CredentialExchange:
+        return credential_record_to_model_v2(record=record)
 
     @classmethod
     def __preview_from_attributes(
         cls,
         attributes: Dict[str, str],
-    ) -> CredentialPreview:
-        return CredentialPreview(
+    ) -> V20CredPreview:
+        return V20CredPreview(
             attributes=[
-                CredAttrSpec(name=name, value=value)
+                V20CredAttrSpec(name=name, value=value)
                 for name, value in attributes.items()
             ]
         )
