+++ app/services/revocation_registry.py
