+++ app/services/trust_registry.py
