+++ app/tests/services/test_trust_registry.py
@@ -3,7 +3,7 @@ from unittest.mock import AsyncMock, Mock
 import pytest
 from httpx import HTTPStatusError, Response
 
-import app.facades.trust_registry as trf
+import app.services.trust_registry as trf
 
 
 @pytest.mark.anyio
